+++ db/migrate/20240703142240_local_authority_main_contact_id.rb
@@ -1,0 +1,5 @@
+class LocalAuthorityMainContactId < ActiveRecord::Migration[7.0]
+  def change
+    add_column :projects, :local_authority_main_contact_id, :uuid
+  end
+end